fix(drag): refactor the droppable placeholder for match-list and drag-in-the-blank PD-1954 PD-1956
