+++ packages/mask-markup/src/choices/index.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import findKey from 'lodash/findKey';
 import Choice from './choice';
-import DroppablePlaceholder from '@pie-lib/drag/lib/droppable-placeholder';
+import DragDroppablePlaceholder from '@pie-lib/drag/lib/drag-in-the-blank-dp';
 
 export default class Choices extends React.Component {
   static propTypes = {
@@ -49,11 +49,11 @@ export default class Choices extends React.Component {
 
     return (
       <div style={elementStyle}>
-        <DroppablePlaceholder disabled={disabled}>
+        <DragDroppablePlaceholder disabled={disabled}>
           {filteredChoices.map((c, index) => (
             <Choice key={`${c.value}-${index}`} disabled={disabled} choice={c} />
           ))}
-        </DroppablePlaceholder>
+        </DragDroppablePlaceholder>
       </div>
     );
   }
